backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis[hiredis]==5.2.0
arq==0.26.1
httpx==0.28.0
python-jose[cryptography]==3.3.0
pyjwt==2.10.0
python-multipart==0.0.17
aiofiles==24.1.0
PyMuPDF==1.24.13
psycopg2-binary==2.9.10
